requirements.txt
streamlit>=1.32.0
playwright>=1.44.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-genai>=1.0.0
anthropic>=0.25.0
openai>=1.23.0
pandas>=2.2.0
openpyxl>=3.1.0
lxml>=5.0.0
